# fix(DB/Creature): Fixed Gordunni Shamans casting too much healing