<!-- First of all, THANK YOU for your contribution. -->
Thanks to the video of @Efymer we could see the real blizzlike logic for those Gordunni.
As such i tried replicating it as close as i could.
I changed the timings and values of the HP missing to get a blizzlike logic.

## Changes Proposed:
-  Increased the cd of the self healing from 14 to 21 seconds to 1:20 to 1:30 minutes.
-  Increased the Hp missing of the friendly creatures for the spell to cast

## Issues Addressed:
<!-- If your fix has a relating issue, link it below -->
- Closes https://github.com/azerothcore/azerothcore-wotlk/issues/7572
- Closes https://github.com/chromiecraft/chromiecraft/issues/1548

## SOURCE:
<!-- If you can, include a source that can strengthen your claim -->
https://www.youtube.com/watch?v=XiqwoaPoksw

## Tests Performed:
<!-- Does it build without errors? Did you test in-game? What did you test? On which OS did you test? Describe any other tests performed -->
- .go c id 5236 
- Attack it. When lower 50% hp he should cast healing wave on himself. CD of 1;20 to 1;30 seconds
- Attack a friendly creature near him. When the creature is at less than 25% hp the shaman should cast the spell and heal him. Cd of 20 seconds.


## How to Test the Changes:
<!-- Describe in a detailed step-by-step order how to test the changes -->

1.  .go c id 5236 
2. Attack it. When lower 50% hp he should cast healing wave on himself. CD of 1;20 to 1;30 seconds
3. Attack a friendly creature near him. When the creature is at less than 25% hp the shaman should cast the spell and heal him. Cd of 20 seconds.

## Known Issues and TODO List:
<!-- Is there anything else left to do after this PR? -->

- Sometimes the friendly heal is not instant. If i remove the timerMin and make it instant the problem is that between 0 to 20 seconds will be cast again. So i have to put it between 20 to 25 seconds.

Test video

https://streamable.com/onno28


<!-- If you intend to contribute repeatedly to our project, it is a good idea to join our discord channel. We set ranks for our contributors and give them access to special resources or knowledge: https://discord.com/invite/DasJqPba)
     Do not remove the instructions below about testing, they will help users to test your PR -->
## How to Test AzerothCore PRs
 
When a PR is ready to be tested, it will be marked as **[WAITING TO BE TESTED]**.

You can help by testing PRs and writing your feedback here on the PR's page on GitHub. Follow the instructions here:

http://www.azerothcore.org/wiki/How-to-test-a-PR

**REMEMBER**: when testing a PR that changes something **generic** (i.e. a part of code that handles more than one specific thing), the tester should not only check that the PR does its job (e.g. fixing spell XXX) but **especially** check that the PR does not cause any regression (i.e. introducing new bugs).

**For example**: if a PR fixes spell X by changing a part of code that handles spells X, Y, and Z, we should not only test X, but **we should test Y and Z as well**.