@syssneck 2021-08-31
Okay, I tested it for a bit and overall I'm happy with the changes. I think these changes come pretty close, to what Efymer recorded on the TBC PTR server.
No superfast double heals anymore:

https://user-images.githubusercontent.com/68868567/131550345-268cfcd0-b4ed-4d80-a5df-692c83d8edc2.mp4

It can be merged from my side.